Organized codebase into modules
