+++ src/lib/JANA/Engine/JEventSourceArrow.h
@@ -5,10 +5,11 @@
 #ifndef JANA2_JEVENTSOURCEARROW_H
 #define JANA2_JEVENTSOURCEARROW_H
 
-#include <JANA/JArrow.h>
-#include <JANA/JMailbox.h>
 #include <JANA/JEvent.h>
 #include <JANA/JFactorySet.h>
+
+#include <JANA/Engine/JArrow.h>
+#include <JANA/Engine/JMailbox.h>
 
 using Event = std::shared_ptr<JEvent>;
 using EventQueue = JMailbox<Event>;
